# docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  models.AddArticleForm:
    properties:
      content:
        type: string
      coverImageUrl:
        type: string
      createdBy:
        type: string
      desc:
        type: string
      state:
        type: integer
      tagID:
        type: integer
      title:
        type: string
    type: object
  models.AddTagForm:
    properties:
      createdBy:
        type: string
      name:
        type: string
      state:
        type: integer
    type: object
  models.EditArticleForm:
    properties:
      content:
        type: string
      coverImageUrl:
        type: string
      desc:
        type: string
      modifled_by:
        type: string
      state:
        type: integer
      tagID:
        type: integer
      title:
        type: string
    type: object
  models.EditTagForm:
    properties:
      modifiedBy:
        type: string
      name:
        type: string
      state:
        type: integer
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license:
    name: MIT
    url: http://github.com/ss75710541/gin-blog/blob/master/LICENSE
  termsOfService: http://github.com/ss75710541/gin-blog
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: TagID
        in: query
        name: tag_id
        type: integer
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        type: string
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 获取多个文章
      tags:
      - 文章
    post:
      parameters:
      - description: Add article
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/models.AddArticleForm'
          type: object
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 新增文章
      tags:
      - 文章
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 删除文章
      tags:
      - 文章
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 获取单个文章
      tags:
      - 文章
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Edit article
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/models.EditArticleForm'
          type: object
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 修改文章
      tags:
      - 文章
  /api/v1/tags:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 获取多个文章标签
      tags:
      - 标签
    post:
      parameters:
      - description: Add tag
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/models.AddTagForm'
          type: object
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 新增文章标签
      tags:
      - 标签
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 删除文章标签
      tags:
      - 标签
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Edit tag
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/models.EditTagForm'
          type: object
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 修改文章标签
      tags:
      - 标签
  /auth:
    get:
      parameters:
      - description: userName
        in: query
        name: username
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Get Auth
      tags:
      - Auth
  /upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image File
        in: formData
        name: image
        required: true
        type: file
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: 上传图片
      tags:
      - Upload
swagger: "2.0"